% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yoctopuce.R
\docType{package}
\name{rYoctoPuceAPI-package}
\alias{rYoctoPuceAPI}
\alias{rYoctoPuceAPI-package}
\title{rYoctoPuceAPI: YoctoPuce USB modules}
\description{
Bridge between R and the yoctopuce Python library, making it possible to control and acquire data from YoctoPuce USB interface modules.
}
\details{
Package 'yoctopuce' provides wrapper functions that call functions
in the 'yoctopuce' Python library through package 'reticulate'.
}
\references{
The documentation of the 'yoctopuce' Python library is installed together
with the library and also available on-line at
\url{https://www.yoctopuce.com/} under the Support menu.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aphalo/rYoctoPuceAPI}
  \item \url{https://docs.r4photobiology.info/rYoctoPuceAPI}
  \item Report bugs at \url{https://github.com/aphalo/rYoctoPuceAPI/issues}
}

}
\author{
\strong{Maintainer}: Pedro J. Aphalo \email{pedro.aphalo@helsinki.fi} (\href{https://orcid.org/0000-0003-3385-972X}{ORCID})

}
